Fixes flakyness.
Really the fix was just the movement of the sleep(.5) to before the
send to proc.stdin call. Its a bit of a mystery why this works and some
other changes didn't, and yes this is potentially fragile, but it works.
